requirements.txt
uWSGI
flask
rq
rq_dashboard
redis
pyTelegramBotAPI
requests
python-dotenv
